+++ Backend_EGM/src/main/java/com/tfg/egm/service/PagoService.java
@@ -3,7 +3,11 @@ package com.tfg.egm.service;
 import com.tfg.egm.entity.Cliente;
 import com.tfg.egm.entity.Pago;
 import com.tfg.egm.repository.PagoRepository;
+
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.lang.StackWalker.Option;
 import java.util.List;
@@ -38,8 +42,15 @@ public class PagoService {
             return pagoRepository.save(pago);
         });
     }
-
+    
     public void deletePago(Long id) {
-        pagoRepository.deleteById(id);
+        if (!pagoRepository.existsById(id)) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "pagoNoExiste");
+        }
+        try {
+            pagoRepository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new ResponseStatusException(HttpStatus.CONFLICT, "pagoNoSePuedeEliminar", e);
+        }
     }
 }
